pyproject.toml
[tool.poetry]
name = "mvp"
version = "0.0.0"
description = "MVP of span classification"
authors = ["Jerry Genser <jerrygenser@gmail.com>"]
packages = [ {include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
spacy = "^3.3"
spacy-transformers = "^1.1.5"
pandas = "^1.4.2"
ipdb = "^0.13"
cupy-cuda116 = "^10.5.0"


[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}
[tool.pytest.ini_options]
addopts = "-ra -q -sv"
asyncio_mode="auto"
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"